+++ src/cuda_extra.cu
@@ -408,7 +408,7 @@ void cryptonight_extra_cpu_prepare(nvid_ctx *ctx, uint32_t startNonce, const xmr
     } else if (algorithm == Algorithm::CN_R) {
         CUDA_CHECK_KERNEL(ctx->device_id, cryptonight_extra_gpu_prepare<Algorithm::CN_R> << <grid, block >> > (wsize, ctx->d_input, ctx->inputlen, startNonce,
             ctx->d_ctx_state, ctx->d_ctx_state2, ctx->d_ctx_a, ctx->d_ctx_b, ctx->d_ctx_key1, ctx->d_ctx_key2));
-    } else if (CnAlgo<>::base(algorithm) == Algorithm::CN_2 || algorithm == Algorithm::CN_PICO_0 || algorithm == Algorithm::CN_PICO_TLO) {
+    } else if (algorithm.base() == Algorithm::CN_2 || algorithm == Algorithm::CN_PICO_0 || algorithm == Algorithm::CN_PICO_TLO) {
         CUDA_CHECK_KERNEL(ctx->device_id, cryptonight_extra_gpu_prepare<Algorithm::CN_2><<<grid, block >>>(wsize, ctx->d_input, ctx->inputlen, startNonce,
             ctx->d_ctx_state, ctx->d_ctx_state2, ctx->d_ctx_a, ctx->d_ctx_b, ctx->d_ctx_key1, ctx->d_ctx_key2));
     } else {
@@ -665,7 +665,7 @@ int cuda_get_deviceinfo(nvid_ctx *ctx)
 
         ctx->device_threads = ((limitedMemory / perThread) / ctx->device_blocks) & 0xFFFFFFFE;
 
-        if (CnAlgo<>::base(ctx->algorithm) == Algorithm::CN_2 && props.major < 6) {
+        if (ctx->algorithm.base() == Algorithm::CN_2 && props.major < 6) {
             // 4 based on my test maybe it must be adjusted later
             size_t threads = 4;
             // 8 is chosen by checking the occupancy calculator
@@ -686,5 +686,5 @@ int cuda_get_deviceinfo(nvid_ctx *ctx)
 
 int cryptonight_gpu_init(nvid_ctx *ctx)
 {
-    return cryptonight_extra_cpu_init(ctx, ctx->algorithm, xmrig_cuda::CnAlgo<>::memory(ctx->algorithm));
+    return cryptonight_extra_cpu_init(ctx, ctx->algorithm, ctx->algorithm.l3());
 }